event_handler.py renamed to getters_setters.py and wrong answer text getters/setters added
